requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pgvector==0.3.6
pydantic==2.10.3
httpx==0.28.1
sentence-transformers==3.3.1
pdfplumber==0.11.4
python-docx==1.1.2
python-multipart==0.0.20
